pyproject.toml
[project]
name = "whisperx-fastapi"
version = "0.1.0"
description = "FastAPI service for WhisperX"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "colorlog==6.9.0",
    "fastapi==0.115.12",
    "gunicorn==23.0.0",
    "httpx==0.28.1",
    "numba==0.61.0",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "tqdm==4.67.1",
    "uvicorn==0.34.3",
    "whisperx==3.3.4",
    "temporalio==1.6.0",
    "psutil==5.9.8",
    "setuptools==70.2.0",
    "torch==2.7.1",
    "torchaudio==2.7.1",
    "torchvision==0.22.1",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.13",
    "aiosignal==1.3.2",
    "alembic==1.16.2",
    "annotated-types==0.7.0",
    "antlr4-python3-runtime==4.9.3",
    "anyio==4.9.0",
    "asteroid-filterbanks==0.4.0",
    "attrs==25.3.0",
    "av==14.4.0",
    "certifi==2025.6.15",
    "cffi==1.17.1",
    "cfgv==3.4.0",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "codespell==2.4.1",
    "coloredlogs==15.0.1",
    "contourpy==1.3.2",
    "coverage==7.9.1",
    "ctranslate2==4.4.0",
    "cycler==0.12.1",
    "distlib==0.3.9",
    "docopt==0.6.2",
    "einops==0.8.1",
    "faster-whisper==1.1.1",
    "filelock==3.13.1",
    "flatbuffers==25.2.10",
    "fonttools==4.58.4",
    "frozenlist==1.7.0",
    "fsspec==2024.6.1",
    "greenlet==3.2.3",
    "h11==0.16.0",
    "hf-xet==1.1.4",
    "httpcore==1.0.9",
    "huggingface-hub==0.33.0",
    "humanfriendly==10.0",
    "hyperpyyaml==1.2.2",
    "identify==2.6.12",
    "idna==3.10",
    "iniconfig==2.1.0",
    "jinja2==3.1.4",
    "joblib==1.5.1",
    "julius==0.2.7",
    "kiwisolver==1.4.8",
    "lightning==2.5.2",
    "lightning-utilities==0.14.3",
    "llvmlite==0.44.0",
    "mako==1.3.10",
    "markdown-it-py==3.0.0",
    "markupsafe==2.1.5",
    "matplotlib==3.10.3",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "multidict==6.5.0",
    "networkx==3.3",
    "nltk==3.9.1",
    "nodeenv==1.9.1",
    "numpy==2.1.2",
    "omegaconf==2.3.0",
    "onnxruntime==1.22.1",
    "optuna==4.4.0",
    "packaging==24.2",
    "pandas==2.3.0",
    "pillow==11.0.0",
    "platformdirs==4.3.8",
    "pluggy==1.6.0",
    "pre-commit==4.2.0",
    "primepy==1.3",
    "propcache==0.3.2",
    "protobuf==6.31.1",
    "psycopg2-binary==2.9.10",
    "pyannote-audio==3.3.2",
    "pyannote-core==5.0.0",
    "pyannote-database==5.1.3",
    "pyannote-metrics==3.2.1",
    "pyannote-pipeline==3.0.1",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pygments==2.19.1",
    "pyparsing==3.2.3",
    "pytest==8.4.0",
    "pytest-cov==6.1.1",
    "python-dateutil==2.9.0.post0",
    "pytorch-lightning==2.5.2",
    "pytorch-metric-learning==2.8.1",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "regex==2024.11.6",
    "requests==2.32.4",
    "rich==14.0.0",
    "ruamel-yaml==0.18.14",
    "ruamel-yaml-clib==0.2.12",
    "ruff==0.11.13",
    "safetensors==0.5.3",
    "scikit-learn==1.7.0",
    "scipy==1.15.3",
    "semver==3.0.4",
    "sentencepiece==0.2.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "soundfile==0.13.1",
    "speechbrain==1.0.3",
    "sqlalchemy==2.0.41",
    "starlette==0.46.2",
    "sympy==1.13.3",
    "tabulate==0.9.0",
    "tensorboardx==2.6.4",
    "threadpoolctl==3.6.0",
    "tokenizers==0.21.0",
    "torch-audiomentations==0.12.0",
    "torch-pitch-shift==1.2.5",
    "torchmetrics==1.7.3",
    "transformers==4.48.0",
    "librosa==0.10.2",
    "typer==0.16.0",
    "types-protobuf==6.30.2.20250703",
    "typing-inspection==0.4.1",
    "typing-extensions==4.12.2",
    "tzdata==2025.2",
    "urllib3==2.5.0",
    "virtualenv==20.31.2",
    "yarl==1.20.1",
]

[project.optional-dependencies]
dev = [
    "codespell==2.4.1",
    "pre-commit==4.2.0",
    "pytest==8.4.0",
    "pytest-cov==6.1.1",
    "ruff==0.11.13",
]
gpu = [
    "nvidia-cublas-cu12==12.6.4.1",
    "nvidia-cuda-cupti-cu12==12.6.80",
    "nvidia-cuda-nvrtc-cu12==12.6.77",
    "nvidia-cuda-runtime-cu12==12.6.77",
    "nvidia-cudnn-cu12==9.5.1.17",
    "nvidia-cufft-cu12==11.3.0.4",
    "nvidia-cufile-cu12==1.11.1.6",
    "nvidia-curand-cu12==10.3.7.77",
    "nvidia-cusolver-cu12==11.7.1.2",
    "nvidia-cusparse-cu12==12.5.4.2",
    "nvidia-cusparselt-cu12==0.6.3",
    "nvidia-nccl-cu12==2.26.2",
    "nvidia-nvjitlink-cu12==12.6.85",
    "nvidia-nvtx-cu12==12.6.77",
    "triton==3.3.1",
]

[tool.isort]
profile = "black"
skip_glob = ["migrations/*"]

[tool.codespell]
skip = ["*.json","*.yaml"]
builtin = "clear,rare,informal,usage,code,names"

[tool.coverage.run]
source = ["app"]
omit = [
    "tests/*",
    "/tmp/*",
    "venv/*",
    "*/virtualenv/*",
    "*/site-packages/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError"
]
ignore_errors = true

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["notebooks*"]
